# swagger.json
{"openapi":"3.1.0","info":{"title":"Spring API","description":"Spring Boot REST API","version":"3.0"},"servers":[{"url":"http://localhost:8080","description":"HTTPS server"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/user/user":{"post":{"tags":["user-controller"],"summary":"Create a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["authentication-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponse"}}}}}}},"/api/v1/auth/signin":{"post":{"tags":["authentication-controller"],"operationId":"signin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthenticationResponse"}}}}}}},"/api/v1/user/{id}":{"get":{"tags":["user-controller"],"summary":"delete One user","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/user/{id}/user":{"get":{"tags":["user-controller"],"summary":"Get one User","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/user/users":{"get":{"tags":["user-controller"],"summary":"find all users","operationId":"findAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}}},"/api/v1/resource":{"get":{"tags":["authorization-controller"],"operationId":"sayHello","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"User":{"properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"roles":{"type":"array","items":{"type":"string"},"uniqueItems":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"}}},"SignUpRequest":{"properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"JwtAuthenticationResponse":{"properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}}},"SigninRequest":{"properties":{"email":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}